#1478: rose edit: better handling of file page.
The page is now a normal rose edit page with a special source
valuewidget.
