+++ demo/rose-config-edit/demo_meta/app/12-file/rose-app.conf
@@ -19,6 +19,9 @@ source=$SDFN
 [file:directory/]
 source=/some/directory/
 
+[file:ignored_source]
+!source=ignored
+
 [file:unexpected_content]
 something=something not right
