+++ demo/rose-config-edit/demo_meta/rose-suite.conf
@@ -1,0 +1,2 @@
+[file:fooey]
+source=/opt/foo/fooey
