+++ etc/rose-meta/rose-app-conf/rose-meta.conf
@@ -33,19 +33,20 @@ description=File installation configuration
 # Note: the following 'file:*' syntax is unique, and has no wildcard.
 # Wildcard operators in metadata sections are not supported.
 [file:*=checksum]
-description=Checksum for external resource (optional)
+description=Checksum for validation
 pattern=^[0-9a-f]{32}$|^$
 
 # Note: the following 'file:*' syntax is unique, and has no wildcard.
 # Wildcard operators in metadata sections are not supported.
 [file:*=mode]
-description=Mode to apply when running - e.g copy the file
+description=Operation to perform
 values=auto, mkdir, symlink
 
 # Note: the following 'file:*' syntax is unique, and has no wildcard.
 # Wildcard operators in metadata sections are not supported.
 [file:*=source]
 description=External resource URI
+sort-key=00
 type=file
 
 [poll]
