+++ etc/rose-meta/rose-suite-conf/rose-meta.conf
@@ -56,6 +56,7 @@ values=auto, mkdir, symlink
 # Wildcard operators in metadata sections are not supported.
 [file:*=source]
 description=External resource URI
+sort-key=00
 type=file
 
 [jinja2:suite.rc]
